+++ DiscordBot/Handlers/ReactionHandler.cs
@@ -8,7 +8,6 @@ using Discord.WebSocket;
 using DiscordBot.Common;
 using DiscordBot.Extensions;
 using DiscordBot.Other;
-using DiscordBot.Logging;
 
 namespace DiscordBot.Handlers
 {
@@ -28,12 +27,6 @@ namespace DiscordBot.Handlers
             if (QuoteHandler.RequestQuoteMessages.Contains(message.Id))
             {
                 await HandleRequestQuoteReactions(message, channel, reaction);
-                return;
-            }
-
-            if (TransactionLogger.TransactionMessages.Contains(message.Id))
-            {
-                await HandleTransactionReactions(message, channel, reaction);
                 return;
             }
 
@@ -133,54 +126,6 @@ namespace DiscordBot.Handlers
                 await message.Value.AddReactionAsync(Extensions.Extensions.ArrowRight);
             }
             else if (QuoteHandler.RequestPageNumber[QuoteHandler.RequestQuoteMessages.IndexOf(message.Id)] == QuoteHandler.GetRequestQuotesListLength)
-            {
-                await message.Value.AddReactionAsync(Extensions.Extensions.ArrowLeft);
-            }
-            else
-            {
-                await message.Value.AddReactionAsync(Extensions.Extensions.ArrowLeft);
-                await message.Value.AddReactionAsync(Extensions.Extensions.ArrowRight);
-            }
-        }
-
-        private static async Task HandleTransactionReactions(Cacheable<IUserMessage, ulong> message, ISocketMessageChannel channel, SocketReaction reaction)
-        {
-            // Check to see if the next page or previous page was clicked.
-            if (reaction.Emote.Name == Extensions.Extensions.ArrowLeft.Name)
-            {
-                if (TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)] == 1)
-                    return;
-
-                TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)]--;
-            }
-            else if (reaction.Emote.Name == Extensions.Extensions.ArrowRight.Name)
-            {
-                if (TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)] == TransactionLogger.GetSplicedTransactonListCount)
-                    return;
-
-                TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)]++;
-            }
-
-            StringBuilder sb = new StringBuilder()
-                .Append("**Transactions**\n**----------------**\n`Total Transactions: " + TransactionLogger.TransactionsList.Count + "`\n```");
-
-            List<string> transactions = TransactionLogger.GetSplicedTransactions(TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)]);
-
-            for (int i = 0; i < transactions.Count; i++)
-            {
-                sb.Append(((i + 1) + ((TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)] - 1) * 10)) + ": " + transactions[i] + "\n");
-            }
-
-            sb.Append("``` `Page " + TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)] + "`");
-
-            await message.Value.ModifyAsync(msg => msg.Content = sb.ToString());
-            await message.Value.RemoveAllReactionsAsync();
-
-            if (TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)] == 1)
-            {
-                await message.Value.AddReactionAsync(Extensions.Extensions.ArrowRight);
-            }
-            else if (TransactionLogger.PageNumber[TransactionLogger.TransactionMessages.IndexOf(message.Id)] == TransactionLogger.GetSplicedTransactonListCount)
             {
                 await message.Value.AddReactionAsync(Extensions.Extensions.ArrowLeft);
             }