+++ DiscordBot/Modules/Public/FunModule.cs
@@ -8,7 +8,6 @@ using DiscordBot.Common;
 using DiscordBot.Common.Preconditions;
 using DiscordBot.Extensions;
 using DiscordBot.Other;
-using DiscordBot.Logging;
 
 namespace DiscordBot.Modules.Public
 {